+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.5] - 16/May/2020
+
+- Payload for ChatMessage is now working as expected.
+- Added [allowReadReceipts] feature. Set to false if you need to reduce message read counts on Firestore.
+
 ## [0.0.4] - 11/May/2020
 
 - Fixed bug for read counts.